修复类型提升导致的子图报错 (#2790)
* fix plt bm env, test=model

* fix plt bm env 1, test=model

* fix plt bm env 2, test=model

* fix plt bm env 3, test=model

* fix plt bm env 4, test=model

* add plt bm module, test=model

* add plt bm module 1, test=model

* add plt bm module 2, test=model

* add plt bm module 2, test=model

* add plt bm module 2, test=model

* add plt bm module 2, test=model

* add plt bm module 2, test=model

* add plt bm module 2, test=model

* add multi plt run, test=model

* add multi plt run, test=model

* add multi plt run, test=model

* add multi plt run, test=model

* add multi plt run, test=model

* add multi plt run, test=model

* add plt multi process, test=model

* add plt multi process, test=model

* add plt multi process, test=model

* add plt multi process, test=model

* add plt multi process, test=model

* add plt multi process, test=model

* add plt multi process, test=model

* fix plt start, test=model

* fix plt start, test=model

* add plt device id for multi, test=model

* add plt device id for multi, test=model

* add plt device id for multi, test=model

* add plt device id for multi, test=model

* add plt device id for multi, test=model

* fix plt get data, test=model

* fix plt get data, test=model

* [CINN] fix plt perf_monitor layercase, test=model

* [CINN] fix plt eval_bm, test=model

* [CINN] fix plt eval_bm, test=model

* fix plt eval.py, test=model

* fix plt eval.py, test=model

* CINN plt train update, test=model

* CINN plt train update, test=model

* CINN plt train update, test=model

* CINN plt train update, test=model

* CINN plt train update, test=model

* update plt yaml, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* update plt compare strategy, test=model

* plt update report, test=model

* plt update report, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add layer api case, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt binary search, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt competition frame, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api layer, test=model

* add plt api nn case, test=model

* add plt api nn case, test=model

* add plt api nn case, test=model

* add plt api nn case, test=model

* add plt api nn case, test=model

* add plt api nn case, test=model

* fix plt api math case, test=model

* fix plt math layer and yml2layer, test=model

* fix plt math layer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* add plt end2end sublayer, test=model

* fix perf plt run, test=model

* fix perf plt run, test=model

* fix perf plt run, test=model

* add plt pts callback, test=model

* add plt pts callback, test=model

* fix plt perf __init__, test=model

* fix plt perf run, test=model

* fix plt perf run, test=model

* fix plt ignore yml, test=model

* update plt dy inputspec, test=model

* update plt dy inputspec, test=model

* update plt dy inputspec, test=model

* update plt dy inputspec, test=model

* 修复类型提升导致的子图报错, test=model

* 修复类型提升导致的子图报错, test=model

+++ framework/e2e/PaddleLT_new/layercase/sublayer1000/Det_cases/ppyoloe_voc_ppyoloe_plus_crn_l_30e_voc/SIR_182.py
@@ -50,7 +50,7 @@ class LayerCase(paddle.nn.Layer):
 def create_tensor_inputs():
     inputs = (
         paddle.rand(shape=[1, 1, 4116], dtype=paddle.float32),
-        paddle.randint(low=0, high=10, shape=[1, 1], dtype=paddle.int32),
+        paddle.randint(low=0, high=10, shape=[1, 1], dtype=paddle.int64),
         paddle.randint(low=0, high=10, shape=[1, 1, 1], dtype=paddle.int32),
         paddle.rand(shape=[1, 4116], dtype=paddle.float32),
         paddle.rand(shape=[1, 1, 4], dtype=paddle.float32),
@@ -63,7 +63,7 @@ def create_tensor_inputs():
 def create_numpy_inputs():
     inputs = (
         np.random.random(size=[1, 1, 4116]).astype('float32'),
-        np.random.randint(low=0, high=10, size=[1, 1], dtype='int32'),
+        np.random.randint(low=0, high=10, size=[1, 1], dtype='int64'),
         np.random.randint(low=0, high=10, size=[1, 1, 1], dtype='int32'),
         np.random.random(size=[1, 4116]).astype('float32'),
         np.random.random(size=[1, 1, 4]).astype('float32'),
